refactor: move dapp example into components into individual directories
