+++ dapp/components/AssetActionsTab/index.ts
@@ -1,0 +1,1 @@
+export { default } from './AssetActionsTab';
