+++ dapp/components/AtomicTransactionActionsTab/AtomicTransactionActionsTab.tsx
@@ -56,7 +56,7 @@ import { theme } from '@extension/theme';
 // types
 import { INetwork } from '@extension/types';
 import { IWindow } from '@external/types';
-import { IAccountInformation, IAssetInformation } from './types';
+import { IAccountInformation, IAssetInformation } from '../../types';
 
 // utils
 import convertToAtomicUnit from '@common/utils/convertToAtomicUnit';
@@ -66,7 +66,7 @@ import {
   createAppCallTransaction,
   createAssetTransferTransaction,
   createPaymentTransaction,
-} from './utils';
+} from '../../utils';
 
 interface IAssetValue extends IAssetInformation {
   amount: BigNumber;
